refactor: split completion in separate files by responsibility
