+++ src/unsafe/providers/completion/index.ts
@@ -1,0 +1,1 @@
+export { doCompletion } from "./completion";
